requirements.txt
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.115.6
filelock==3.20.2
fsspec==2025.12.0
future==1.0.0
grpcio==1.76.0
grpcio-tools==1.76.0
h11==0.14.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==0.17.3
httplib2==0.31.2
httptools==0.7.1
httpx==0.24.1
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
mpmath==1.3.0
mysql-connector-python==9.1.0
nebula3-python==3.8.2
networkx==3.6.1
numpy==2.4.0
packaging==25.0
pillow==12.1.0
pluggy==1.6.0
portalocker==2.10.1
protobuf==6.33.2
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
pyparsing==3.3.2
pytest==7.4.2
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.12.1
regex==2025.11.3
requests==2.32.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==2.7.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.32.1
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
